+++ Specs/Scene/CompositePrimitiveSpec.js
@@ -96,12 +96,12 @@ defineSuite([
     function createLabels(position) {
         position = position || { x : -1.0, y : 0.0, z : 0.0 };
         var labels = new LabelCollection();
+        labels.clampToPixel = false;
         labels.add({
             position : position,
             text : 'x',
             horizontalOrigin : HorizontalOrigin.CENTER,
-            verticalOrigin : VerticalOrigin.CENTER,
-            font : '1px sans-serif'
+            verticalOrigin : VerticalOrigin.CENTER
         });
         return labels;
     }
@@ -863,6 +863,7 @@ defineSuite([
         // collection to a Cesium app and looking for when it is/is not occluded.
 
         var labels = new LabelCollection();
+        labels.clampToPixel = false;
         labels.add({
             position : Ellipsoid.WGS84.cartographicToCartesian(new Cartographic.fromDegrees(-75.10, 39.57)),
             text : 'x',